requirements.txt
fastapi
requests
python-dotenv
phidata
yfinance
duckduckgo_search
uvicorn
groq